move database analyzing to the database folder
